don't allow zooming out until the zoom has started
Our fork of zoom.js preloads the image before zooming it in. This
is necessary because we don't know what image file to zoom until
the user clicks.

Some things break if `close` gets called before `zoomOriginal`
gets called. `close` expects the image to be in a "zoomed" state.

I briefly tried to make `close` more reliable no matter what state
the zoomer is in, but I don't think it's worth it. It's also an
odd user experience to "stop loading" but clicking twice.

Instead, just don't let `close` get called until the preload
finishes.

Also, there is a rare case where `transitionend` won't fire if
you zoom out right when the "zoom in" animation starts. Add
a setTimeout fallback for this case.